Modularize :media module
