--- features/media/src/main/kotlin/app/ss/media/playback/model/PlaybackProgressState.kt
@@ -1,19 +1,0 @@
-package app.ss.media.playback.model
-
-import androidx.compose.runtime.Immutable
-import app.ss.media.playback.extensions.millisToDuration
-
-@Immutable
-data class PlaybackProgressState(
-    val total: Long = 0L,
-    val position: Long = 0L,
-    val elapsed: Long = 0L,
-    val buffered: Long = 0L
-) {
-
-    val progress get() = ((position.toFloat() + elapsed) / (total + 1).toFloat()).coerceIn(0f, 1f)
-    val bufferedProgress get() = ((buffered.toFloat()) / (total + 1).toFloat()).coerceIn(0f, 1f)
-
-    val currentDuration get() = (position + elapsed).millisToDuration()
-    val totalDuration get() = total.millisToDuration()
-}
