+++ libraries/media/model/src/main/kotlin/ss/libraries/media/model/PlaybackQueue.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2023. Adventech <info@adventech.io>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package ss.libraries.media.model
+
+import app.ss.models.media.AudioFile
+
+data class PlaybackQueue(
+    val list: List<String> = emptyList(),
+    val audiosList: List<AudioFile> = emptyList(),
+    val title: String? = null,
+    val initialMediaId: String = "",
+    val currentIndex: Int = 0
+) : List<AudioFile> by audiosList {
+
+    val isValid = list.isNotEmpty() && audiosList.isNotEmpty() && currentIndex >= 0
+
+    val currentAudio get() = getOrNull(currentIndex)
+}
